requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
PyJWT
python-dotenv
slowapi
